+++ .github/changelog.md
@@ -10,3 +10,9 @@
 - **Branche** : feature/readme
 - **Message** : Mise à jour du Readme et du script pr.mjs 
 
+## 2025-05-15
+
+- **Branche** : feature/electron
+- **Message** : Mise à jour de l'integration electron 
+- **Dernière modification par** : Liam Von Astoria <liam.dispa@skygenesisenterprise.com>
+